delete logs

+++ backend/main.py
@@ -165,24 +165,24 @@ async def retrieve_entries(user_id: int, db: Session = Depends(get_database)):
 
 
 
-@app.get('/delete_entry')
-async def delete_entry(entry_id, db: Session = Depends(get_database)):
+@app.post('/delete_entry')
+async def delete_entry(entry_id: int, db: Session = Depends(get_database)):
     try:
-        entry = db.query(Log).filter(Log.id == entry_id)
-        print(entry)
+        entry = db.query(Log).filter(Log.id == entry_id).first()
 
-        deleted_entry = DeletedLog()
-        deleted_entry.turbidity = entry.turbidity
-        deleted_entry.ph = entry.ph
-        deleted_entry.tds = entry.tds
-        deleted_entry.ec = entry.ec
-        deleted_entry.battery = entry.battery
-        deleted_entry.date_created = entry.date_created
-        deleted_entry.record_owner = entry.record_owner
+        if entry:
+            deleted_entry = DeletedLog()
+            deleted_entry.turbidity = entry.turbidity
+            deleted_entry.ph = entry.ph
+            deleted_entry.tds = entry.tds
+            deleted_entry.ec = entry.ec
+            deleted_entry.battery = entry.battery
+            deleted_entry.date_created = entry.date_created
+            deleted_entry.record_owner = entry.record_owner
 
-        db.add(deleted_entry)
-        db.delete(entry)
-        db.commit()
+            db.add(deleted_entry)
+            db.delete(entry)
+            db.commit()
 
         return { 'response': 'Data Deleted', 'status_code': 200 }
     except:
